- refactored Summary, deleted Summary factory - refactored Controller endpoints
