+++ build.gradle
@@ -64,7 +64,7 @@ tasks.jacocoTestCoverageVerification {
 	violationRules {
 		rule {
 			limit {
-				minimum = "0.52".toBigDecimal()
+				minimum = "0.40".toBigDecimal()
 			}
 		}
 	}
