+++ src/main/java/com/tw/personalfinancetracker/model/dto/Summary.java
@@ -2,7 +2,13 @@ package com.tw.personalfinancetracker.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.tw.personalfinancetracker.model.Transaction;
 import lombok.*;
+
+import java.util.List;
+
+import static com.tw.personalfinancetracker.util.Constants.EXPENSE;
+import static com.tw.personalfinancetracker.util.Constants.INCOME;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -15,4 +21,22 @@ public class Summary {
     private Double totalIncome;
     private Double totalExpense;
     private Double balance;
+
+    public Summary(List<Transaction> transactions) {
+        if (!transactions.isEmpty()) {
+            Double totalIncome = sumAllTransactions(transactions, INCOME);
+            Double totalExpense = sumAllTransactions(transactions, EXPENSE);
+
+            this.totalIncome = totalIncome;
+            this.totalExpense = totalExpense;
+            this.balance = totalIncome - totalExpense;
+        }
+    }
+
+    private static Double sumAllTransactions(List<Transaction> transactions, String transactionType) {
+        return transactions.stream()
+                .filter(t -> transactionType.equals(t.getType()))
+                .map(Transaction::getAmount)
+                .reduce(0.0, Double::sum);
+    }
 }
